requirements.txt
flask
pandas
requests
Pillow==8.0.0